--- a/context.go
+++ b/context.go
@@ -7,13 +7,13 @@ package fox
 import (
 	"context"
 	"fmt"
+	"github.com/tigerwill90/fox/internal/netutil"
 	"io"
 	"iter"
 	"net"
 	"net/http"
 	"net/url"
 	"slices"
-	"strings"
 )
 
 // ContextCloser extends [Context] for manually created instances, adding a Close method
@@ -169,7 +169,7 @@ func (c *cTx) SetWriter(w ResponseWriter) {
 func (c *cTx) RemoteIP() *net.IPAddr {
 	ipStr, _, _ := net.SplitHostPort(c.req.RemoteAddr)
 
-	ip, zone := splitHostZone(ipStr)
+	ip, zone := netutil.SplitHostZone(ipStr)
 	ipAddr := &net.IPAddr{
 		IP:   net.ParseIP(ip),
 		Zone: zone,
@@ -413,17 +413,4 @@ func WrapH(h http.Handler) HandlerFunc {
 
 		h.ServeHTTP(c.Writer(), c.Request())
 	}
-}
-
-func splitHostZone(s string) (host, zone string) {
-	// This is copied from an unexported function in the Go stdlib:
-	// https://github.com/golang/go/blob/5c9b6e8e63e012513b1cb1a4a08ff23dec4137a1/src/net/ipsock.go#L219-L228
-
-	// The IPv6 scoped addressing zone identifier starts after the last percent sign.
-	if i := strings.LastIndexByte(s, '%'); i > 0 {
-		host, zone = s[:i], s[i+1:]
-	} else {
-		host = s
-	}
-	return
 }